feat: Add Redux module for campaigns
This commit introduces a new Redux module to manage campaigns.

The new module includes:
- A 'campaigns' slice with actions and reducers to handle state updates for campaign data.
- Sagas to manage asynchronous API calls for fetching and updating campaign information.
- Type definitions for campaigns, API payloads, and action payloads, consolidated in the entities folder.

The root reducer and root saga have been updated to integrate the new campaigns module.
File naming conventions have been updated based on user feedback.
